Move app subcommands out of main.go
This way, subcommands don't show up on platforms on which they're not supported.
